Adding No Content View with animations

+++ TodoList/Views/ListView.swift
@@ -12,24 +12,32 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
-                    }
-               
-            }.onDelete(perform: listViewModel.deleteItem)
-                .onMove(perform: listViewModel.moveItem)
-            
-        }.navigationTitle("Todo 📝")
+        
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
+                       
+                    }.onDelete(perform: listViewModel.deleteItem)
+                        .onMove(perform: listViewModel.moveItem)
+                    
+                }.listStyle(.plain)
+            }
+        }
+        .navigationTitle("Todo 📝")
             .navigationBarItems(
                 leading: EditButton(),
                 trailing: NavigationLink("Add", destination: AddView())
             )
-            .listStyle(.plain)
     }
     
 }
